requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
streamlit==1.28.2
bandit==1.7.5
semgrep==1.51.0
checkov==3.1.25
boto3==1.33.0
azure-identity==1.15.0
azure-mgmt-storage==21.1.0
azure-mgmt-containerservice==27.0.0
azure-mgmt-authorization==3.0.0
reportlab==4.0.7
pandas==2.1.3
altair==5.1.2
pyyaml==6.0.1
requests==2.31.0